Reworked The Dragon Radar Code Slightly.
I was told I should "Use +y instead of -y" to locate blocks.
I also tried using ModTags, but it didn't work after working on it for 30 minutes so I just hard coded it...

On top of that, I fixed a unintentional bug where all Valuable Blocks were found as the same type.
Also minor changes to it allows us to dynamically find Dragon Balls, or print unknown blocks as a Debug!